Revert "rename error-config.ts & error-config.spec.ts"
This reverts commit be926de9b1a6bc8952935b97058c8d18e99fe4d0.
